requirements.txt
langchain
pinecone-client
openai
python-dotenv
tqdm
PyPDF2
python-docx